requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-auth==2.23.2
google-auth-oauthlib==1.1.0
requests==2.31.0
colorama>=0.4.4
flask-sqlalchemy==3.0.5
flask-login==0.6.3
python-dotenv==1.0.0
PyJWT==2.8.0